refactor: automatically select the optimal packet backend
